Add files via upload
complete

+++ 13.quickNest/controller/userController.js
@@ -1,18 +1,17 @@
-import cloudinary from "../config/cloudinary.js";
 import HttpError from "../middleware/HttpError.js";
-import uploads from "../middleware/upload.js";
-import User from "../model/userModel.js";
+import User from "../model/User.js";
+import cloudinary from "../config/cloudinary.js";
 
 const add = async (req, res, next) => {
   try {
-    const { name, email, password, phone, role } = req.body;
+    const { name, email, password, role, phone } = req.body;
 
     const newUser = {
       name,
       email,
       password,
-      phone,
       role,
+      phone,
       profilePic: req.file ? req.file.path : "undefined",
       cloudinaryId: req.file ? req.file.filename : "undefined",
     };
@@ -25,7 +24,7 @@ const add = async (req, res, next) => {
 
     res.status(201).json({ success: true, user });
   } catch (error) {
-    next(error);
+    next(new HttpError(error.message, 500));
   }
 };
 
@@ -38,14 +37,12 @@ const login = async (req, res, next) => {
     const token = await user.generateAuthToken();
 
     if (!user) {
-      return next(new HttpError("Unable to login ", 400));
+      return next(new HttpError("unable to login"));
     }
 
-    res
-      .status(200)
-      .json({ success: true, message: "Login Successfully", user, token });
+    res.status(200).json({ success: true, user, token });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
@@ -54,29 +51,26 @@ const authLogin = async (req, res, next) => {
     const user = req.user;
 
     if (!user) {
-      return next(new HttpError("Unable to login"));
+      return next(new HttpError("user not found", 404));
     }
 
-    res.status(201).json({ success: true, user });
+    res.status(200).json({ success: true, user });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
 const logOut = async (req, res, next) => {
   try {
-    const token = req.token; 
-
     req.user.tokens = req.user.tokens.filter((t) => {
-      return t.token !== token;
+      return t.token != req.token;
     });
 
     await req.user.save();
 
-    res.status(200).json({
-      success: true,
-      message: "User LogOut Successfully",
-    });
+    res
+      .status(200)
+      .json({ success: true, message: "user logOut successfully" });
   } catch (error) {
     next(new HttpError(error.message, 500));
   }
@@ -88,33 +82,36 @@ const logOutAll = async (req, res, next) => {
 
     await req.user.save();
 
-    res
-      .status(200)
-      .json({ success: true, message: "User Logout from all device" });
+    res.status(200).json({
+      success: true,
+      message: "user logOut from all device successfully",
+    });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
-const getAll = async (req, res, next) => {
+const allUser = async (req, res, next) => {
   try {
     const users = await User.find({});
 
     if (users.length === 0) {
-      res
-        .status(200)
-        .json({ success: true, message: "No any users are there" });
+      res.status(200).json({ success: true, message: "no user data found" });
     }
 
-    res.status(200).json({ success: true, message: "all users", users });
+    res
+      .status(200)
+      .json({ success: true, message: "all user data fetched", users });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
 const update = async (req, res, next) => {
   try {
-    const user = req.user;
+    let targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser);
 
     if (!user) {
       return next(new HttpError("user not found", 404));
@@ -122,18 +119,32 @@ const update = async (req, res, next) => {
 
     const updates = Object.keys(req.body);
 
-    const allowedField = ["name", "password", "phone"];
+    let allowedFields = ["name", "password", "phone", "profilePic"];
 
-    const isValid = updates.every((field) => allowedField.includes(field));
+    if (req.user.role === "admin" || req.user.role === "super_admin") {
+      allowedFields = [...allowedFields, "role", "isVerified"];
+    }
+
+    const isValid = updates.every((field) => allowedFields.includes(field));
 
     if (!isValid) {
-      return next(new HttpError("only allowed fields can be updated", 400));
+      return next(new HttpError("only allowed field can be updated", 400));
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
     }
 
     updates.forEach((update) => (user[update] = req.body[update]));
 
     if (req.file) {
-      await cloudinary.uploader.destroy(user.cloudinaryId);
+      if (user.cloudinaryId) {
+        await cloudinary.uploader.destroy(user.cloudinaryId);
+      }
 
       user.profilePic = req.file.path;
 
@@ -144,23 +155,39 @@ const update = async (req, res, next) => {
 
     res
       .status(200)
-      .json({ success: true, message: "user update successfully", user });
+      .json({ success: true, message: "user Data updated successfully", user });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message));
   }
 };
 
 const deleteUser = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser);
+
+    if (!user) {
+      return next(new HttpError("user not found"), 401);
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
+    }
 
     await User.deleteOne(user);
 
-    await cloudinary.uploader.destroy(user.cloudinaryId);
+    if (user.cloudinaryId) {
+      await cloudinary.uploader.destroy(user.cloudinaryId);
+    }
 
     res
       .status(200)
-      .json({ success: true, message: "user delete successfully" });
+      .json({ success: true, message: "user deleted successfully" });
   } catch (error) {
     next(new HttpError(error.message, 500));
   }
@@ -172,7 +199,7 @@ export default {
   authLogin,
   logOut,
   logOutAll,
-  getAll,
+  allUser,
   update,
   deleteUser,
 };
